Final fix for sub-package import setup
