--- whispercpp_runner/__init__.py
@@ -1,8 +1,0 @@
-"""This File initializes the runners for the transcription process"""
-from whispercpp_runner.src.runner import Runner
-
-RUNNER_COUNT = 1
-# Make it possible to run the app directly from the command line
-
-runner = Runner(1, "large-v1")
-runner.run()
